==> 3_visualise.Rmd <==
---
title: "3_visualise"
author: "Fiona Neilson"
date: "27/06/2021"
output:
  word_document: default
  html_document: default
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# 3 VISUALISE DATA

## 3.1 Import libraries

```{r, message=FALSE, warning=FALSE}
library(ggplot2)
install.packages("treemapify")
library(treemapify) 
install.packages("ggpubr")
library(ggpubr)
library(dplyr)
library(forcats)
library(tidyverse)
```

## 3.2 Read in dataframe

```{r}
df = read.csv("out_2_df.csv", header=TRUE)
```

## 3.3 Create treemap 

I chose a treemap to show the proportions of languages at the ABS 1DC level, which is the highest level language group name. I selected this to give an overall view of the balance of languages rather than a precise breakdown. 

Joining the languages classification to my dataframe earlier made it possible for me to link individual languages back to their parent groups.

```{r, fig.cap = "Fig 1"}
# create treemap
df %>% count(LANP, GroupName1DC, wt = total) %>% ggplot(aes(area = n, fill = GroupName1DC, label = LANP)) +  geom_treemap() +  geom_treemap_text(fontface = "italic", colour = "white", place = "centre",  grow = TRUE) + labs(fill = "Language Group") + ggtitle("Languages other than English spoken in the \nCity of Greater Dandenong")
```

## 3.4 Create bar chart

I created a chart to show the level of education and English proficiency by language group, to look for any distinctive features.

### 3.4.1 Reorder levels for HEAP 

The ordered factors that I created in the first file were dropped from the dataframe when I exported and then re-imported the data as a CSV file, as part of my recreation of the dataframe in inRStudio Cloud. I re-ordered some of these to ensure the correct display in plots.

```{r}
# reorder levels
education <- c("Non_Secondary", "Secondary", "Higher") # create vector in correct order
education <- as.factor(education)
df$HEAP <- factor(df$HEAP, levels = education, ordered = TRUE)
levels(df$HEAP)
```
### 3.4.2 Create plot 

```{r, fig.cap = "Fig 2"}
# create bar chart 
g21 <- mutate(group_by(df, HEAP, GroupName1DC), Prop = total / sum(total))
ggplot(g21) + geom_col(aes(GroupName1DC, Prop, fill = ENGP), position = "stack") + facet_wrap(~HEAP, ncol = 4) + coord_flip() + theme(axis.text.x = element_text(angle = 90, hjust = 1)) + labs(fill = "English") + xlab("Language Group") + ylab("Proportion")
```

## 3.5 Create faceted scatterplots

I used faceted scatterplots to explore the relationship between language, education, English proficiency and gender within a sample language group. While this is the maximum number of variables I would normally include on one chart, this is a useful exploratory tool that provides a multi-faceted summary. I repeated this for every 1DC language group. 

```{r, fig.cap = "Fig 3"}
# create faceted scatterplots 
df %>% filter(GroupName1DC == "Eastern Asian Languages") %>% ggplot +  geom_point(mapping = aes(x = HEAP, y = SEXP, colour = ENGP, size = total), position = position_jitter(width = 0.2, height = 0.2),  alpha = 0.6) + labs(colour = "English", size = "Count") + xlab("Educational level") + ylab("Gender") + facet_wrap(.~LANP, ncol = 4) + theme(axis.text.x = element_text(angle = 45, hjust = 1))
```

## 3.6 Explore socially advantaged/ disadvantaged groups

### 3.6.1 Create dataframe for disadvantaged groups

I created criteria for social advantage and disadvantage to explore the characteristics of these groups. While the choice of criteria is subjective and is not informed by any particular method, what I aim to show here is the ability to create criteria and then carry out analysis of a subset. 


#### 3.6.1.1 Combine year levels into two groups

I combined the single year 2016 into the 2006-2015 band as exploration showed no value in leaving this year as a standalone level. 

```{r}
#combine levels
df <- mutate(df, YARRP = fct_recode(YARRP, "2006-16" = "2016"))
df <- mutate(df, YARRP = fct_recode(YARRP, "2006-16" = "2006-15"))
head(df, 25)
```

#### 3.6.1.2 Reorder levels chronologically

```{r}
# order levels: YARRP 
years2 <- c("Pre_2006", "2006-16") # create vector in correct order
years2 <- as.factor(years2)
df$YARRP <- factor(df$YARRP, levels = years2, ordered = TRUE)
levels(df$YARRP)
```

#### 3.6.1.3 Apply disadvantage criteria to create a subset

```{r}
# create disadvantaged df
Disadvan <- df %>% filter(ENGP == "Inadequate", NEDD == "No_Internet", EETP == "Not_Engaged", HEAP == "Non_Secondary")
Disadvan %>% group_by(LANP, BPLP, SEXP) %>% count(LANP, wt = total) %>% arrange(desc(n))
# create a vector of the languages with min 30 count
DisadvanLang <- Disadvan %>% count(LANP, wt = total) %>% filter(n >= 30)
```

```{r}
levels(Disadvan$YARRP)
```

#### 3.6.1.4 Create graph of languages with minimum count of 30 

I removed languages with a total of less than 30 as sample sizes below this threshold are not statistically representative. 

```{r, fig.cap = "Fig 4"}
# create graph of just the langs with >= 30 
Disadvan %>% subset(LANP %in% DisadvanLang$LANP) %>%   ggplot(aes(x=LANP, y=total, fill=SEXP)) + geom_bar(stat="identity", position="dodge") + coord_flip() + facet_grid(.~YARRP) + theme(legend.position = "top",  axis.text.x = element_text(angle = 45, hjust = 1, vjust = 0.5)) +  xlab("Language") + ylab("Count >= 30") + labs(fill = "Gender")
```

#### 3.6.1.5 Exploration code

I created this code to explore the disadvantaged group by level; just insert the relevant variable.

```{r}
# use this code to explore the counts by sex
Disadvan %>% count(SEXP, wt = total) 
```
```{r}
# compute summary 
Disadvanperc <- sum(Disadvan$total)/sum(df$total) * 100    
round(Disadvanperc, 2)
```

```{r}
Disadvan %>% count(LANP, wt = total) %>% arrange(desc(n))
```


```{r}
sum(Disadvan$total) 
```


### 3.6.2 Create dataframe for advantaged groups

```{r}
# create advantaged df
Advan <- df %>% filter(ENGP == "Adequate", NEDD == "Internet", EETP == "Fully", HEAP == "Higher") 
```

#### 3.6.2.1 Compute sum and percentage

```{r}
Advanperc_M <- 6063/ (6063 + 4053)
Advanperc_F <- 4053 / (6063 + 4053) 
round(Advanperc_M, 2)
round(Advanperc_F, 2)
sum(df$total) 
sum(Advan$total) 
```
## 3.7 Plot advantaged groups

### 3.7.1 Create a vector of languages with minimum count of 30 

```{r}
AdvanLang <- Advan %>% count(LANP, wt = total) %>%   filter(n >= 30) %>% arrange(desc(n))
AdvanLang # This computes 40 languages
```

### 3.7.3 Separate languages into groups of 10 for neater plotting 

```{r}
# "Binning' the languages into 4 groups by rank
OneTen <- AdvanLang[1:10,] #top 10
ElevenTwen <- AdvanLang[11:20,]
TwenOneThir <- AdvanLang[21:30,]
ThirOneForty <- AdvanLang[31:40,]
```

### 3.7.4 Recode a long name

```{r}
# recode long name Advan <- Advan %>% mutate(LANP = fct_recode(LANP,  "Persian" = "Persian (excluding Dari)"))
TwenOneThir <- TwenOneThir %>% mutate(LANP = fct_recode(LANP,  "Persian" = "Persian (excluding Dari)"))
```

### 3.7.5 Create plots for each language 'bin' group

```{r}
# 1 – 10
a1 <- Advan %>% subset(LANP %in% OneTen$LANP) %>%   ggplot(aes(x=LANP, y=total, fill=SEXP)) + geom_bar(stat="identity", position="dodge") + coord_flip() + facet_grid(.~YARRP) + ylab("") + theme(legend.position="none") + xlab(NULL) 

# 11 – 20
a2 <- Advan %>% subset(LANP %in% ElevenTwen$LANP) %>%   ggplot(aes(x=LANP, y=total, fill=SEXP)) + geom_bar(stat="identity", position="dodge") + coord_flip() + facet_grid(.~YARRP) + ylab("") + theme(legend.position="none") + xlab(NULL) 

# 21 – 30
a3 <- Advan %>% subset(LANP %in% TwenOneThir$LANP) %>% ggplot(aes(x=LANP, y=total, fill=SEXP)) +  geom_bar(stat="identity", position="dodge") + coord_flip() +  facet_grid(.~YARRP) +  ylab("") + theme(legend.position="none") + xlab(NULL) 

# 31 – 40
a4 <- Advan %>% subset(LANP %in% ThirOneForty$LANP) %>%   ggplot(aes(x=LANP, y=total, fill=SEXP)) + geom_bar(stat="identity", position="dodge") + coord_flip() +  facet_grid(.~YARRP) + ylab("") + xlab(NULL) + labs(fill = "Gender")
```

### 3.7.6 Display as a panel

```{r, fig.cap = "Fig 5"}
ggarrange(a1, a2, a3, a4, ncol=2, nrow=2, common.legend = TRUE, legend="bottom")
```
